refactor: TalkPickSection 컴포넌트에 대한 컴포넌트명 및 폴더명 수정
