+++ src/components/organisms/TalkPickSection/TalkPickSection.style.ts
